+++ src/main/kotlin/dev/nyon/bbm/extensions/Networking.kt
@@ -1,6 +1,5 @@
 package dev.nyon.bbm.extensions
 
-import dev.nyon.bbm.config.serverConfig
 import net.minecraft.network.protocol.common.custom.CustomPacketPayload
 import net.minecraft.server.level.ServerPlayer
 
@@ -8,5 +7,5 @@ fun sendToClient(player: ServerPlayer, packet: CustomPacketPayload) {
     //? if fabric
     net.fabricmc.fabric.api.networking.v1.ServerPlayNetworking.send(player, packet)
     //? if neoforge
-    /*net.neoforged.neoforge.network.PacketDistributor.sendToPlayer(player, serverConfig!!)*/
+    /*net.neoforged.neoforge.network.PacketDistributor.sendToPlayer(player, packet)*/
 }