+++ changelog.md
@@ -1,1 +1,3 @@
-- fix error in console on NeoForge, when JumpCollisionPacket should have been sent, but Config packet was sent
+- fixed a bug when serverConfig was wrongfully set on client startup
+    - this led to the player being able to use the mod's features even though the server did not explicitly permit it to
+      do so